pyproject.toml
[tool.poetry]
name = "lichessanalyser"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
networkx = "^2.5"
python-chess = "^1.999"
discord = "^1.0.1"
keepalive = "^0.5"
flask = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
